+++ docs/storybook/CHANGELOG.md
@@ -1,5 +1,110 @@
 # @livekit/component-docs-storybook
 
+## 1.0.40
+
+### Patch Changes
+
+- Updated dependencies [[`f9b3d53cb345a12a89dbd3ae7bf4264018e6ed89`](https://github.com/livekit/components-js/commit/f9b3d53cb345a12a89dbd3ae7bf4264018e6ed89)]:
+  - @livekit/components-react@2.0.6
+
+## 1.0.39
+
+### Patch Changes
+
+- Updated dependencies [[`6c49a946246a2361f406d40f814465da9ae87dce`](https://github.com/livekit/components-js/commit/6c49a946246a2361f406d40f814465da9ae87dce)]:
+  - @livekit/components-react@2.0.5
+
+## 1.0.38
+
+### Patch Changes
+
+- Update livekit-client to 2.0.10 for Chrome 124 compatibility - [#806](https://github.com/livekit/components-js/pull/806) ([@davidzhao](https://github.com/davidzhao))
+
+- Updated dependencies [[`614ba929468d0de124374f50a584412f53ece56b`](https://github.com/livekit/components-js/commit/614ba929468d0de124374f50a584412f53ece56b), [`24de53696cd6c90cf32f0e35d3b64d6b20d8dc5b`](https://github.com/livekit/components-js/commit/24de53696cd6c90cf32f0e35d3b64d6b20d8dc5b), [`30882e2b55652511945dbf5d6a55fe54f7c4f158`](https://github.com/livekit/components-js/commit/30882e2b55652511945dbf5d6a55fe54f7c4f158)]:
+  - @livekit/components-react@2.0.4
+  - @livekit/components-styles@1.0.11
+
+## 1.0.37
+
+### Patch Changes
+
+- Updated dependencies [[`cd79814c344a6271d5cd5b415037c60e3393b52d`](https://github.com/livekit/components-js/commit/cd79814c344a6271d5cd5b415037c60e3393b52d), [`ef3a674b2ff8e0bd2ce325704c6100f359f9bcf0`](https://github.com/livekit/components-js/commit/ef3a674b2ff8e0bd2ce325704c6100f359f9bcf0)]:
+  - @livekit/components-react@2.0.3
+
+## 1.0.36
+
+### Patch Changes
+
+- Updated dependencies [[`fee3db5d2eedeeb0ccb5aa4b81509a6550607b89`](https://github.com/livekit/components-js/commit/fee3db5d2eedeeb0ccb5aa4b81509a6550607b89)]:
+  - @livekit/components-react@2.0.2
+
+## 1.0.35
+
+### Patch Changes
+
+- Updated dependencies [[`ffda1ad09c951b1a86bed94b74dc3b0c2fb4a7a3`](https://github.com/livekit/components-js/commit/ffda1ad09c951b1a86bed94b74dc3b0c2fb4a7a3), [`344dd70e23c7a1184cab6c066d5cf6ef9c919eb2`](https://github.com/livekit/components-js/commit/344dd70e23c7a1184cab6c066d5cf6ef9c919eb2)]:
+  - @livekit/components-react@2.0.1
+  - @livekit/components-styles@1.0.10
+
+## 1.0.34
+
+### Patch Changes
+
+- Updated dependencies [[`6d3e25c593923eb86ee66b76455bb55a7da779b5`](https://github.com/livekit/components-js/commit/6d3e25c593923eb86ee66b76455bb55a7da779b5), [`db00029d104b034b9af1a091d653da242de0753a`](https://github.com/livekit/components-js/commit/db00029d104b034b9af1a091d653da242de0753a), [`e00d930a80bbcba30933656c484e064a683c5408`](https://github.com/livekit/components-js/commit/e00d930a80bbcba30933656c484e064a683c5408), [`39fb8ce298e74c655551afa48afd58e511fba550`](https://github.com/livekit/components-js/commit/39fb8ce298e74c655551afa48afd58e511fba550), [`5fa2d6b33c591a5cf511c015d066546c983bf2aa`](https://github.com/livekit/components-js/commit/5fa2d6b33c591a5cf511c015d066546c983bf2aa), [`52215f13191f56895d564513ac0bb7edb2de8e60`](https://github.com/livekit/components-js/commit/52215f13191f56895d564513ac0bb7edb2de8e60)]:
+  - @livekit/components-react@2.0.0
+
+## 1.0.33
+
+### Patch Changes
+
+- Updated dependencies [[`f041101c`](https://github.com/livekit/components-js/commit/f041101ce2808b2d431f27c929af8a574c7b9b19), [`ce2cc799`](https://github.com/livekit/components-js/commit/ce2cc799c33e158699ef173ff4e9d6927b206a23), [`21edb909`](https://github.com/livekit/components-js/commit/21edb909f9844b345772d10620d351cfc903bed9)]:
+  - @livekit/components-react@1.5.3
+  - @livekit/components-styles@1.0.9
+
+## 1.0.32
+
+### Patch Changes
+
+- Updated dependencies [[`6ea54f87`](https://github.com/livekit/components-js/commit/6ea54f87ace371cd5611372e1485d8ad82d200f1), [`ba28381a`](https://github.com/livekit/components-js/commit/ba28381a00006c059f6f5d0be72a1e03276cf7f0), [`d3bb268c`](https://github.com/livekit/components-js/commit/d3bb268c9e3e5f27ae3d55f45c01f7c3bbca1f91), [`3e15bde4`](https://github.com/livekit/components-js/commit/3e15bde4720309869dc8f6b42ea7263f1bc8319e)]:
+  - @livekit/components-react@1.5.2
+
+## 1.0.31
+
+### Patch Changes
+
+- Updated dependencies [[`dab1e274`](https://github.com/livekit/components-js/commit/dab1e274dc41a83c97dc7a415b8bd69d888c7da9)]:
+  - @livekit/components-react@1.5.1
+
+## 1.0.30
+
+### Patch Changes
+
+- Updated dependencies [[`3c6564fd`](https://github.com/livekit/components-js/commit/3c6564fdaeabbe6cb676595e410f4bbfbc9e9614), [`dfdd550b`](https://github.com/livekit/components-js/commit/dfdd550bc11a592ee45b545892e2ec9c29f7e6e9)]:
+  - @livekit/components-react@1.5.0
+  - @livekit/components-styles@1.0.8
+
+## 1.0.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @livekit/components-react@1.4.2
+
+## 1.0.28
+
+### Patch Changes
+
+- Updated dependencies [[`00be594b`](https://github.com/livekit/components-js/commit/00be594b8d42e4f391311499d7ba7b003533cb32), [`da5ffe3b`](https://github.com/livekit/components-js/commit/da5ffe3b867de4d8567a544597905dd9b2fc8fd9), [`adc3d042`](https://github.com/livekit/components-js/commit/adc3d0427745000b73b70971297e9496c7eee9d6), [`11ec05d4`](https://github.com/livekit/components-js/commit/11ec05d4a753cd4e317e5031ccdb66137cab49b5), [`2dab3b08`](https://github.com/livekit/components-js/commit/2dab3b0855b09eb8c72c4c0c6fc63de5d5f6384d), [`57229c69`](https://github.com/livekit/components-js/commit/57229c697a7f0b0be52f328e7115cf7bbefd7665)]:
+  - @livekit/components-react@1.4.1
+
+## 1.0.27
+
+### Patch Changes
+
+- Updated dependencies [[`d0b90cba`](https://github.com/livekit/components-js/commit/d0b90cbae07abf701c9c1de403152e0aafb915a0), [`01611f5d`](https://github.com/livekit/components-js/commit/01611f5d2498c8dcdc231e5e305f281c9c2d9311), [`946d2b06`](https://github.com/livekit/components-js/commit/946d2b06d78a5f8ab721a98fbebe6b16a2e98463), [`66df197a`](https://github.com/livekit/components-js/commit/66df197a861429f6bb87774ec5be4cad1bd5762a), [`d52a53b5`](https://github.com/livekit/components-js/commit/d52a53b5727c9cceccd7050dbae737d1fc47acb8), [`8348816d`](https://github.com/livekit/components-js/commit/8348816d83a943535efe90827eca4bb56a1948a6), [`67d62ba3`](https://github.com/livekit/components-js/commit/67d62ba3cb902d5b55fb553c1ab8480a12fb1d55), [`01b04180`](https://github.com/livekit/components-js/commit/01b0418077da58c99fc3c81e76a46dd739517011), [`8d760ead`](https://github.com/livekit/components-js/commit/8d760ead2aa7101e608773df04ce09e9e9b5967a)]:
+  - @livekit/components-react@1.4.0
+  - @livekit/components-styles@1.0.7
+
 ## 1.0.26
 
 ### Patch Changes